refactor: moved parsing code into map module
