+++ src/index.js
@@ -1,76 +1,10 @@
-import choroplethMap from './choroplethMap'
+import { choroplethMap, build_driving_map, build_races_run_map, parseRaces } from './choroplethMap'
 
 const margin = { left: 120, right: 300, top: 20, bottom: 120 };
 
 const visualization = d3.select('#visualization');
 const visualizationDiv = visualization.node();
 const svg = visualization.select('svg');
-
-/***** parsing code for choropleth *********/
-
-const drivingTimesMap = {};
-const build_driving_map = row => {
-  drivingTimesMap[row.Town] = {};
-  drivingTimesMap[row.Town].time = +row.DrivingTime;
-  const hours = Math.floor(+row.DrivingTime/60);
-  const mins = +row.DrivingTime - 60*hours;
-  if(hours > 0) {
-    drivingTimesMap[row.Town].timeString = hours + "h " + mins + " min";
-  } else {
-    drivingTimesMap[row.Town].timeString = mins + " min";
-  }
-  if(!(row.Town in raceHorizonByTown)) {
-    raceHorizonByTown[row.Town] = { 'daysToRace': 400, 'raceType': ""};
-  }
-  return row;
-};
-
-const racesRunMap = {};
-const build_races_run_map = row => {
-  racesRunMap[row.Town] = {};
-  racesRunMap[row.Town].distance = row.Distance;
-  return row;
-};
-
-const today = d3.timeDay(new Date());
-const racesSoonByTown = {};
-const raceHorizonByTown = {};
-const fmt = d3.format("02");
-const parseRaces = row => {
-  row.Month = +row.Month;
-  row.Day = +row.Day;
-  row.Weekday = +row.Weekday;
-  row.DateString = fmt(row.Month) + "/" + fmt(row.Day);
-  row.raceDay = d3.timeDay(new Date(2017, row.Month-1, row.Day));
-  const daysToRace = d3.timeDay.count(today, row.raceDay);
-  if(daysToRace >= 0 && daysToRace <= 14) {
-    const raceString = "<tr><td><span class='racedate'>" + 
-          row["Date/Time"] + 
-          "</span></td><td><span class='racedistance'>" + 
-          row.Distance + "</span></td><td><span class='racename'>" + 
-          row.Name + "</span></td></tr>";          
-    if(row.Town in racesSoonByTown) {
-      racesSoonByTown[row.Town] += raceString;
-    } else {
-      racesSoonByTown[row.Town] = "<table>" + raceString;
-    }
-    const raceType = daysToRace <= 7 ? "hasRaceVerySoon" : "hasRaceSoon"; 
-    if(row.Town in raceHorizonByTown) {
-      if(daysToRace < raceHorizonByTown[row.Town].daysToRace) {
-        raceHorizonByTown[row.Town] = { 
-          'daysToRace': daysToRace, 
-          'raceType': raceType 
-        };            
-      }            
-    } else {
-      raceHorizonByTown[row.Town] = { 
-        'daysToRace': daysToRace, 
-        'raceType': raceType 
-      };            
-    }
-  }
-  return row;
-};
 
 const functions = {
   calendar: choroplethMap,
@@ -155,11 +89,7 @@ function dataLoaded(error, mapData, drivingTimes, racesRun, races) {
         mapData,
         drivingTimes,
         racesRun,
-        races,
-        racesRunMap, 
-        drivingTimesMap, 
-        racesSoonByTown, 
-        raceHorizonByTown
+        races
       ],
       margin: margin
     },
@@ -168,11 +98,7 @@ function dataLoaded(error, mapData, drivingTimes, racesRun, races) {
         mapData,
         drivingTimes,
         racesRun,
-        races,
-        racesRunMap, 
-        drivingTimesMap, 
-        racesSoonByTown, 
-        raceHorizonByTown
+        races
       ],
       margin: margin
     },
@@ -181,11 +107,7 @@ function dataLoaded(error, mapData, drivingTimes, racesRun, races) {
         mapData,
         drivingTimes,
         racesRun,
-        races,
-        racesRunMap, 
-        drivingTimesMap, 
-        racesSoonByTown, 
-        raceHorizonByTown
+        races
       ],
       margin: margin
     },
@@ -194,11 +116,7 @@ function dataLoaded(error, mapData, drivingTimes, racesRun, races) {
         mapData,
         drivingTimes,
         racesRun,
-        races,
-        racesRunMap, 
-        drivingTimesMap, 
-        racesSoonByTown, 
-        raceHorizonByTown
+        races
       ],
       margin: margin
     }
